+++ src/cimon-desktop/src/internal-preload/index.d.ts
@@ -1,0 +1,8 @@
+import { IElectronAPI } from './types';
+
+declare global {
+  interface Window {
+    electronAPI: IElectronAPI;
+  }
+}
+export * from './types';